archive_mysql_tool添加mysql监控定期打印

+++ monitor-agent/archive_mysql_tool/funcs/db.go
@@ -35,7 +35,7 @@ type ConnectionTrendPoint struct {
 	OpenConnections  int
 	InUseConnections int
 	IdleConnections  int
-	WaitCount        int
+	WaitCount        int64
 	WaitDuration     time.Duration
 	UsageRate        float64
 	ReuseRate        float64
